Issue #2394: Move Headers.Common -> Headers.Names & Headers.Values.
This is a breaking change but is clearer and easier to use (less
nesting).
