+++ components/service/pocket/src/main/java/mozilla/components/service/pocket/PocketEndpointRaw.kt
@@ -7,7 +7,7 @@ package mozilla.components.service.pocket
 import android.net.Uri
 import android.support.annotation.WorkerThread
 import mozilla.components.concept.fetch.Client
-import mozilla.components.concept.fetch.Headers.Common.USER_AGENT
+import mozilla.components.concept.fetch.Headers.Names.USER_AGENT
 import mozilla.components.concept.fetch.MutableHeaders
 import mozilla.components.concept.fetch.Request
 import mozilla.components.service.pocket.ext.fetchBodyOrNull
