+++ components/service/pocket/src/test/java/mozilla/components/service/pocket/PocketEndpointRawTest.kt
@@ -5,7 +5,7 @@
 package mozilla.components.service.pocket
 
 import mozilla.components.concept.fetch.Client
-import mozilla.components.concept.fetch.Headers.Common.USER_AGENT
+import mozilla.components.concept.fetch.Headers.Names.USER_AGENT
 import mozilla.components.concept.fetch.Response
 import mozilla.components.service.pocket.helpers.MockResponses
 import mozilla.components.service.pocket.helpers.assertRequestParams
